feat: extract login page background to own component for consistency
