+++ src/routes/Login/Page.svelte
@@ -5,9 +5,8 @@
     import Layout from '../../components/Layout.svelte';
     import TabItem from '../../components/Tab/TabItem.svelte';
     import TabGroup from '../../components/Tab/TabGroup.svelte';
-    import { TOP_BLOB, BOT_BLOB } from './blobPaths';
-    import Blob from '../../components/Blob.svelte';
     import Drippy from '../../assets/drippy-animated.svelte';
+    import Background from './Background.svelte';
     let user = true;
 </script>
 
@@ -22,13 +21,7 @@
             <button class:user>Login</button>
         </form>
     </main>
-
-    <div class="blob top">
-        <Blob ds={TOP_BLOB} duration="20s" />
-    </div>
-    <div class="blob bot">
-        <Blob ds={BOT_BLOB} duration="28s" />
-    </div>
+    <Background />
     <TabGroup>
         <TabItem on:click={() => (user = true)} active={user}>
             <FaUser />
@@ -40,20 +33,6 @@
 </Layout>
 
 <style>
-    .blob {
-        @apply text-blue-300;
-        @apply fixed z-[-999];
-        @apply w-[720px];
-    }
-
-    .top {
-        @apply left-[-420px] top-[-245px];
-    }
-
-    .bot {
-        @apply bottom-[-144px] right-[-420px];
-    }
-
     main {
         @apply wrapper pb-12;
         @apply flex flex-col items-center justify-evenly;
